webalyzeapi/requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
mypy==0.790
mypy-extensions==0.4.3
pytz==2020.1
sqlparse==0.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
